services/inference/requirements.txt

fastapi
uvicorn
joblib
numpy
pillow

torch==2.2.2
torchvision==0.17.2
scikit-learn
